backend/pyproject.toml
[project]
name = "rf4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "gunicorn>=23.0.0",
    "inference-sdk>=0.12.0",
    "ipywidgets>=8.1.6",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.1",
    "numpy>=2.2.5",
    "openai>=1.75.0",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "pymysql>=1.1.1",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.0",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.2",
    "seaborn>=0.13.2",
    "uvicorn>=0.34.2",
    "uvicorn-worker>=0.3.0",
]
